+++ examples/deal.II/CMakeLists.txt
@@ -22,12 +22,12 @@ which conflict with the requirements."
 endif()
 
 DEAL_II_INITIALIZE_CACHED_VARIABLES()
-PROJECT("bp")
+PROJECT("bps")
 
 DEAL_II_INITIALIZE_CACHED_VARIABLES()
 
-ADD_EXECUTABLE(bp bp.cc)
-DEAL_II_SETUP_TARGET(bp)
+ADD_EXECUTABLE(bps bps.cc)
+DEAL_II_SETUP_TARGET(bps)
 
-TARGET_INCLUDE_DIRECTORIES(bp PUBLIC ${CEED_DIR}/include)
-TARGET_LINK_LIBRARIES(bp ${CEED_DIR}/lib/libceed.so)
+TARGET_INCLUDE_DIRECTORIES(bps PUBLIC ${CEED_DIR}/include)
+TARGET_LINK_LIBRARIES(bps ${CEED_DIR}/lib/libceed.so)
