+++ examples/deal.II/bps.cc
@@ -45,7 +45,7 @@
 #include <sstream>
 
 // include operators
-#include "bp.h"
+#include "bps.h"
 
 
 /**
